⚡️ add custom hook to listen realtime updates
